Revert to what we had in athena 8.3 (lprng 3.5.3).
git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@15714 728af825-273c-0410-89f9-f7d3b574a069

+++ third/lprng/lprng/src/common/lpd_remove.c
@@ -1,14 +1,14 @@
 /***************************************************************************
  * LPRng - An Extended Print Spooler System
  *
- * Copyright 1988-2000, Patrick Powell, San Diego, CA
+ * Copyright 1988-1999, Patrick Powell, San Diego, CA
  *     papowell@astart.com
  * See LICENSE for conditions of use.
  *
  ***************************************************************************/
 
  static char *const _id =
-"$Id: lpd_remove.c,v 1.5 2000-04-03 19:01:47 mwhitson Exp $";
+"$Id: lpd_remove.c,v 1.6 2001-03-07 01:19:34 ghudson Exp $";
 
 
 #include "lp.h"
@@ -61,9 +61,9 @@ int Job_remove( int *sock, char *input )
 
 	/* get the options */
 	++input;
-	DEBUGF(DLPRM1)("Job_remove: input '%s'", input );
+	DEBUG1("Job_remove: input '%s'", input );
 	Split(&tokens,input,Whitespace,0,0,0,0,0);
-	DEBUGFC(DLPRM2)Dump_line_list("Job_remove: input", &tokens );
+	if(DEBUGL2)Dump_line_list("Job_remove: input", &tokens );
 
 	/* check printername for characters, underscore, digits */
 
@@ -74,13 +74,13 @@ int Job_remove( int *sock, char *input )
 	}
 	name = tokens.list[0];
 
-	DEBUGF(DLPRM1)("Job_remove: checking '%s'", name );
+	DEBUG1("Job_remove: checking '%s'", name );
 	if( (s = Clean_name( name )) ){
 		plp_snprintf( error, sizeof(error),
 			_("printer '%s' has illegal character at '%s' in name"), name, s );
 		goto error;
 	}
-	DEBUGF(DLPRM1)("Job_remove: result '%s'", name );
+	DEBUG1("Job_remove: result '%s'", name );
 	Set_DYN(&Printer_DYN,name);
 
 	user = safestrdup(tokens.list[1],__FILE__,__LINE__);
@@ -93,8 +93,8 @@ int Job_remove( int *sock, char *input )
 	Check_max(&tokens,1);
 	tokens.list[tokens.count] = 0;
 
-	if( safestrcmp( Printer_DYN, ALL ) ){
-		DEBUGF(DLPRM2)( "Job_remove: checking printcap entry '%s'",  Printer_DYN );
+	if( strcmp( Printer_DYN, ALL ) ){
+		DEBUG2( "Job_remove: checking printcap entry '%s'",  Printer_DYN );
 		Set_DYN(&Printer_DYN, Printer_DYN );
 		Get_queue_remove( user, sock, &tokens, &done_list );
 	} else {
@@ -108,11 +108,11 @@ int Job_remove( int *sock, char *input )
 
  error:
 	logmsg( LOG_INFO, _("Job_remove: error '%s'"), error );
-	DEBUGF(DLPRM2)("Job_remove: error msg '%s'", error );
+	DEBUG2("Job_remove: error msg '%s'", error );
 	safestrncat(error,"\n");
 	if( Write_fd_str( *sock, error ) < 0 ) cleanup(0);
  done:
-	DEBUGF(DLPRM2)( "Job_remove: done" );
+	DEBUG2( "Job_remove: done" );
 	if( user ) free(user); user = 0;
 	Free_line_list(&done_list);
 	Free_line_list(&tokens);
@@ -129,19 +129,18 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 {
 	char msg[SMALLBUFFER], header[SMALLBUFFER];
 	int control_perm, permission, count, removed, status,
-		i, c = 0, pid, fd;
+		i, c = 0, pid, fd, active_pid;
 	char *s, *identifier, *path;
 	struct stat statb;
-	struct line_list info, active_pid;
+	struct line_list info;
 	struct job job;
 
 	Init_line_list(&info);
-	Init_line_list(&active_pid);
 	Init_job(&job);
 
 	/* set printer name and printcap variables */
-	DEBUGFC(DLPRM2)Dump_line_list("Get_queue_remove - tokens", tokens );
-	DEBUGF(DLPRM2)( "Get_queue_remove: user '%s', printer '%s'",
+	if(DEBUGL2)Dump_line_list("Get_queue_remove - tokens", tokens );
+	DEBUG2( "Get_queue_remove: user '%s', printer '%s'",
 		user, Printer_DYN );
 
 	Errorcode = 0;
@@ -151,31 +150,8 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 
 	status = Setup_printer( Printer_DYN, msg, sizeof(msg) );
 	if( status ){
-		DEBUGF(DLPRM2)("Get_queue_remove: no spool queue '%s'", Printer_DYN);
+		DEBUG2("Get_queue_remove: no spool queue '%s'", Printer_DYN);
 		return;
-	}
-
-	c = Debug;
-	i = DbgFlag;
-	s = Find_str_value(&Spool_control,DEBUG,Value_sep);
-	if( !s ) s = New_debug_DYN;
-	Parse_debug( s, 0 );
-
-	if( !(DbgFlag & DLPRMMASK) ){
-		Debug = c;
-		DbgFlag = i;
-	} else {
-		i = Debug;
-		Debug = c;
-		if( Log_file_DYN ){
-			fd = Trim_status_file( Log_file_DYN, Max_log_file_size_DYN,
-				Min_log_file_size_DYN );
-			if( fd > 0 && fd != 2 ){
-				dup2(fd,2);
-				close(fd);
-			}
-		}
-		Debug = i;
 	}
 
 	/* set up status */
@@ -197,8 +173,8 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 	Perm_check.host = 0;
 	Perm_check.user = 0;
 
-	control_perm = Perms_check( &Perm_line_list, &Perm_check, 0, 0 );
-	DEBUGF(DLPRM2)("Job_status: permission '%s'", perm_str(control_perm));
+	control_perm = Perms_check( &Perm_line_list, &Perm_check, 0 );
+	DEBUG2("Job_status: permission '%s'", perm_str(control_perm));
 
 	if( control_perm != P_ACCEPT ) control_perm = 0;
 
@@ -208,19 +184,20 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 
 	Free_line_list( &Sort_order );
 	Scan_queue( Spool_dir_DYN, &Spool_control, &Sort_order,0,0,0,0 );
-	DEBUGF(DLPRM2)("Get_queue_remove: total files %d", Sort_order.count );
+	DEBUG2("Get_queue_remove: total files %d", Sort_order.count );
 
 	/* scan the files to see if there is one which matches */
 	removed = 0;
-	DEBUGFC(DLPRM3)Dump_line_list("Get_queue_remove - tokens", tokens );
+	active_pid = 0;
+	if(DEBUGL3)Dump_line_list("Get_queue_remove - tokens", tokens );
 	for( count = 0; count < Sort_order.count; ++count ){
 		Free_job(&job);
 		s = Sort_order.list[count];
-		if( (s = safestrchr(Sort_order.list[count],';')) ){
+		if( (s = strchr(Sort_order.list[count],';')) ){
 			Split(&job.info,s+1,";",1,Value_sep,1,1,1);
 		}
 
-		DEBUGFC(DLPRM3)Dump_job("Get_queue_remove - info",&job);
+		if(DEBUGL3)Dump_job("Get_queue_remove - info",&job);
         if( tokens->count && Patselect( tokens, &job.info, 0) ){
 			continue;
         }
@@ -230,7 +207,7 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 		if( !identifier ) identifier
 			= Find_str_value(&job.info,TRANSFERNAME,Value_sep);
 
-		DEBUGF(DLPRM3)("Get_queue_remove: matched '%s'", identifier );
+		DEBUG3("Get_queue_remove: matched '%s'", identifier );
 		plp_snprintf( msg, sizeof(msg), _("  checking perms '%s'\n"),
 			identifier );
 		Write_fd_str( *sock, msg );
@@ -245,7 +222,7 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 			Perm_check.host = &PermHost_IP;
 		}
 		Perm_check.service = 'M';
-		permission = Perms_check( &Perm_line_list, &Perm_check, &job, 1 );
+		permission = Perms_check( &Perm_line_list, &Perm_check, &job );
 		if( permission == P_REJECT && control_perm == 0 ){
 			plp_snprintf( msg, sizeof(msg), _("  no permissions '%s'\n"),
 				      identifier );
@@ -257,14 +234,14 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 		if( permission == P_REJECT ) {
 			/* If someone's using control bits and doesn't
 			   have normal permission, log it for posterity. */
-			if( Perm_check.auth_client_id )
-				logmsg(LOG_INFO, "%s (auth): lprm %s:%s", Perm_check.auth_client_id,
+			if( Auth_client_id_DYN )
+				logmsg(LOG_INFO, "%s (auth): lprm %s:%s", Auth_client_id_DYN,
 				       Printer_DYN, identifier);
 			else
 				logmsg(LOG_INFO, "%s@%s (unauth): lprm %s:%s", user,
 				       FQDNRemote_FQDN, Printer_DYN, identifier);
 		}
-		DEBUGF(DLPRM4)("Get_queue_remove: removing '%s'", identifier );
+		DEBUG4("Get_queue_remove: removing '%s'", identifier );
 		plp_snprintf( msg, sizeof(msg), _("  dequeued '%s'\n"),
 			identifier );
 		Write_fd_str( *sock, msg );
@@ -279,11 +256,8 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 		}
 		setmessage( &job, "LPRM", "success" );
 		if( (pid = Find_flag_value(&job.info,SERVER,Value_sep)) ){
-			DEBUGF(DLPRM4)("Get_queue_remove: active_pid %d", pid );
-			if( kill( pid, 0 ) == 0 ){
-				Check_max(&active_pid,1);
-				active_pid.list[active_pid.count++] = (void *)pid;
-			}
+			DEBUG4("Get_queue_remove: killing %d", pid );
+			active_pid = pid;
 		}
 		++removed;
 		if( tokens->count == 0 ) break;
@@ -292,23 +266,20 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 	Free_job(&job);
 	Free_line_list( &Sort_order );
 	if( removed ){
-		for( i = 0; i < active_pid.count; ++i ){
-			pid = (int)(active_pid.list[i]);
-			active_pid.list[i] = 0;
-			DEBUGF(DLPRM2)("Get_queue_remove: killing pid '%d' SIGINT", pid );
-			kill( pid, SIGINT );
-			kill( pid, SIGCONT );
-		}
 		/* kill spooler process */
-		path = Make_pathname( Spool_dir_DYN, Queue_lock_file_DYN );
+		path = Make_pathname( Spool_dir_DYN, Printer_DYN );
 		pid = 0;
 		if( (fd = Checkread( path, &statb ) ) >= 0 ){
 			pid = Read_pid( fd, (char *)0, 0 );
 			close( fd );
 		}
-		DEBUGF(DLPRM2)("Get_queue_status: checking server pid %d", pid );
+		DEBUG2("Get_queue_status: checking server pid %d", pid );
 		free(path);
 		/* kill active spooler */
+		if( active_pid && active_pid != pid ){
+			kill( active_pid, SIGINT );
+			kill( pid, SIGCONT );
+		}
 		if( pid > 0 ){
 			kill( pid, SIGUSR2 );
 			/* sigh ... yes, you may need to start it */
@@ -320,40 +291,32 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 		Free_line_list(&info);
 		Split(&info, Server_names_DYN, File_sep, 0,0,0,0,0);
 		for( i = 0; i < info.count; ++i ){
-			DEBUGF(DLPRM2)("Get_queue_status: getting subserver status '%s'", 
+			DEBUG2("Get_queue_status: getting subserver status '%s'", 
 				info.list[i] );
 			Set_DYN(&Printer_DYN,info.list[i]);
 			Get_local_or_remote_remove( user, sock, tokens, done_list );
-			DEBUGF(DLPRM2)("Get_queue_status: finished subserver status '%s'", 
+			DEBUG2("Get_queue_status: finished subserver status '%s'", 
 				info.list[i] );
 		}
 	} else if( Destinations_DYN ){
 		Free_line_list(&info);
 		Split(&info, Destinations_DYN, File_sep, 0,0,0,0,0);
 		for( i = 0; i < info.count; ++i ){
-			DEBUGF(DLPRM2)("Get_queue_status: getting destination status '%s'", 
+			DEBUG2("Get_queue_status: getting destination status '%s'", 
 				info.list[i] );
 			Set_DYN(&Printer_DYN,info.list[i]);
 			Get_local_or_remote_remove( user, sock, tokens, done_list );
-			DEBUGF(DLPRM2)("Get_queue_status: finished destination status '%s'", 
+			DEBUG2("Get_queue_status: finished destination status '%s'", 
 				info.list[i] );
 		}
 	} else if( Bounce_queue_dest_DYN ){
-		DEBUGF(DLPRM2)("Get_queue_status: getting bouncequeue dest status '%s'", 
+		DEBUG2("Get_queue_status: getting bouncequeue dest status '%s'", 
 			Bounce_queue_dest_DYN);
 		Set_DYN(&Printer_DYN,Bounce_queue_dest_DYN);
 		Get_local_or_remote_remove( user, sock, tokens, done_list );
-		DEBUGF(DLPRM2)("Get_queue_status: finished subserver status '%s'", 
+		DEBUG2("Get_queue_status: finished subserver status '%s'", 
 			Bounce_queue_dest_DYN );
 	} else if( RemoteHost_DYN ){
-		/* put user name at start of list */
-		Check_max(tokens,2);
-		for( i = tokens->count; i > 0; --i ){
-			tokens->list[i] = tokens->list[i-1];
-		}
-		tokens->list[0] = user;
-		++tokens->count;
-		tokens->list[tokens->count] = 0;
 		fd = Send_request( 'M', REQ_REMOVE, tokens->list, Connect_timeout_DYN,
 			Send_query_rw_timeout_DYN, *sock );
 		if( fd >= 0 ){
@@ -362,17 +325,13 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 			}
 			close(fd);
 		}
-		for( i = 0; i < tokens->count; ++i ){
-			tokens->list[i] = tokens->list[i+1];
-		}
-		--tokens->count;
 	}
 
-	DEBUGF(DLPRM2)("Get_queue_remove: finished '%s'", Printer_DYN );
+	DEBUG2("Get_queue_remove: finished '%s'", Printer_DYN );
 	goto done;
 
  error:
-	DEBUGF(DLPRM2)("Get_queue_remove: error msg '%s'", msg );
+	DEBUG2("Get_queue_remove: error msg '%s'", msg );
 	plp_snprintf(header, sizeof(header), "Printer: %s", Printer_DYN );
 	safestrncpy( header, _(" ERROR: ") );
 	safestrncat( header, msg );
@@ -380,7 +339,6 @@ void Get_queue_remove( char *user, int *sock, struct line_list *tokens,
 	Write_fd_str( *sock, header );
  done:
 	Free_line_list(&info);
-	Free_line_list(&active_pid);
 	Free_job(&job);
 	return;
 }
@@ -389,11 +347,11 @@ void Get_local_or_remote_remove( char *user, int *sock,
 	struct line_list *tokens, struct line_list *done_list )
 {
 	char msg[LARGEBUFFER];
-	int fd, n, i;
+	int fd, n;
 
 	/* we have to see if the host is on this machine */
 
-	if( !safestrchr(Printer_DYN,'@') ){
+	if( !strchr(Printer_DYN,'@') ){
 		Get_queue_remove( user, sock, tokens, done_list );
 		return;
 	}
@@ -405,14 +363,6 @@ void Get_local_or_remote_remove( char *user, int *sock,
 		Get_queue_remove( user, sock, tokens, done_list );
 		return;
 	}
-	/* put user name at start of list */
-	Check_max(tokens,2);
-	for( i = tokens->count; i > 0; --i ){
-		tokens->list[i] = tokens->list[i-1];
-	}
-	tokens->list[0] = user;
-	++tokens->count;
-	tokens->list[tokens->count] = 0;
 	fd = Send_request( 'M', REQ_REMOVE, tokens->list, Connect_timeout_DYN,
 		Send_query_rw_timeout_DYN, *sock );
 	if( fd >= 0 ){
@@ -421,10 +371,6 @@ void Get_local_or_remote_remove( char *user, int *sock,
 		}
 		close(fd);
 	}
-	for( i = 0; i < tokens->count; ++i ){
-		tokens->list[i] = tokens->list[i+1];
-	}
-	--tokens->count;
 }
 
 int Remove_file( char *openname )
@@ -432,7 +378,7 @@ int Remove_file( char *openname )
 	struct stat statb;
 	int fail = 0;
 	if( openname ){
-		DEBUGF(DLPRM3)("Remove_file: removing '%s'", openname );
+		DEBUG3("Remove_file: removing '%s'", openname );
 		unlink( openname );
 		if( stat( openname, &statb ) == 0 ){
 			logerr( LOG_INFO, "Remove_file: unlink did not remove '%s'",
@@ -452,15 +398,14 @@ int Remove_job( struct job *job )
 
 	Setup_cf_info( Spool_dir_DYN, 0, job );
 
-	DEBUGFC(DLPRM3)Dump_job("Remove_job",job);
-	setmessage(job,STATE,"REMOVE");
+	if(DEBUGL3)Dump_job("Remove_job",job);
 	identifier = Find_str_value(&job->info,IDENTIFIER,Value_sep);
 	setmessage( job, TRACE, "remove START" );
 	if( !identifier ){
 		identifier = Find_str_value(&job->info,TRANSFERNAME,Value_sep);
 	}
 
-	DEBUGF(DLPRM3)("Remove_job: identifier '%s'",identifier);
+	DEBUG3("Remove_job: identifier '%s'",identifier);
 	fail = 0;
 	for( i = 0; i < job->datafiles.count; ++i ){
 		datafile = (void *)job->datafiles.list[i];
